+++ src/TonemappingOperators/ashikhmin02/pfstmo_ashikhmin02.cpp
@@ -66,8 +66,8 @@ void calculateLuminance(pfs::Array2Df *Y, float &avLum, float &maxLum,
 #ifdef _OPENMP
     #pragma omp for nowait
 #endif
-    for (int y = 0; y < Y->getRows(); ++y) {
-        int x = 0;
+    for (size_t y = 0; y < Y->getRows(); ++y) {
+        size_t x = 0;
 #ifdef __SSE2__
         for (; x < Y->getCols() - 3; x+=4) {
             vfloat Yv = LVFU((*Y)(x, y));
