+++ src/TonemappingOperators/pattanaik00/tmo_pattanaik00.cpp
@@ -491,7 +491,7 @@ float VisualAdaptationModel::calculateLogAvgLuminance(const pfs::Array2Df &Y) {
 
     float avLum = 0.0f;
 
-    int size = Y.getCols() * Y.getRows();
+    size_t size = Y.getCols() * Y.getRows();
 #ifdef _OPENMP
     #pragma omp parallel
 #endif
@@ -504,8 +504,8 @@ float VisualAdaptationModel::calculateLogAvgLuminance(const pfs::Array2Df &Y) {
 #ifdef _OPENMP
     #pragma omp for nowait
 #endif
-    for (int y = 0; y < Y.getRows(); ++y) {
-        int x = 0;
+    for (size_t y = 0; y < Y.getRows(); ++y) {
+        size_t x = 0;
 #ifdef __SSE2__
         for (; x < Y.getCols() - 3; x+=4) {
             avLumThrv += xlogf(LVFU(Y(x,y)) + c1v);
